+++ huaweicloud/services/acceptance/dataarts/resource_huaweicloud_dataarts_architecture_data_standard_test.go
@@ -58,7 +58,7 @@ func getDataStandardResourceFunc(cfg *config.Config, state *terraform.ResourceSt
 		if dataStandard != nil {
 			break
 		}
-		total := utils.PathSearch("data.value.total", getDataStandardRespBody, 0).(float64)
+		total := utils.PathSearch("data.value.total", getDataStandardRespBody, float64(0)).(float64)
 		currentTotal += len(records.([]interface{}))
 		if currentTotal == int(total) {
 			break
