+++ huaweicloud/services/acceptance/gaussdb/resource_huaweicloud_gaussdb_mysql_account_privilege_test.go
@@ -71,7 +71,7 @@ func getGaussDBAccountPrivilegeResourceFunc(cfg *config.Config, state *terraform
 			account = res
 			break
 		}
-		total := utils.PathSearch("total_count", getGaussDBAccountRespBody, 0).(float64)
+		total := utils.PathSearch("total_count", getGaussDBAccountRespBody, float64(0)).(float64)
 		currentTotal += pageNum
 		if currentTotal == int(total) {
 			break
