+++ huaweicloud/services/ddm/data_source_huaweicloud_ddm_schemas.go
@@ -190,7 +190,7 @@ func flattenGetSchemasResponseBodySchema(d *schema.ResourceData, resp interface{
 	if resp == nil {
 		return 0, 0, nil
 	}
-	total = int(utils.PathSearch("total", resp, 0).(float64))
+	total = int(utils.PathSearch("total", resp, float64(0)).(float64))
 	curJson := utils.PathSearch("databases", resp, make([]interface{}, 0))
 	curArray := curJson.([]interface{})
 	rst = make([]interface{}, 0, len(curArray))
