+++ huaweicloud/services/gaussdb/resource_huaweicloud_gaussdb_mysql_account.go
@@ -228,7 +228,7 @@ func resourceGaussDBAccountRead(_ context.Context, d *schema.ResourceData, meta 
 			account = res
 			break
 		}
-		total := utils.PathSearch("total_count", getGaussDBAccountRespBody, 0).(float64)
+		total := utils.PathSearch("total_count", getGaussDBAccountRespBody, float64(0)).(float64)
 		currentTotal += pageNum
 		if currentTotal == int(total) {
 			break
