+++ huaweicloud/services/gaussdb/resource_huaweicloud_gaussdb_mysql_database.go
@@ -220,7 +220,7 @@ func resourceGaussDBDatabaseRead(_ context.Context, d *schema.ResourceData, meta
 			database = db
 			break
 		}
-		total := utils.PathSearch("total_count", getGaussDBDatabaseRespBody, 0).(float64)
+		total := utils.PathSearch("total_count", getGaussDBDatabaseRespBody, float64(0)).(float64)
 		currentTotal += pageNum
 		if currentTotal == int(total) {
 			break
